change file size to long instead (since there are some huge numbers involved)

+++ service/src/main/java/eu/ebrains/kg/search/model/source/openMINDSv3/FileV3.java
@@ -42,7 +42,7 @@ public class FileV3 extends SourceInstanceV3 {
     @Getter
     @Setter
     public static class Size {
-        private int value;
+        private Long value;
         private String unit;
     }
